Add files via upload
Implemented a pair trading strategies on stocks in Indian Banking stocks. The strategies focused on correlation between the stocks and implemented pair on a combination of banking stocks. The code can be used to test pair trading in different sectors, training between the period 2010-2020 and then testing the performance between 2020-2022.